refactor: modernize MicMute application architecture
- Replaced manual memory management with Modern C++ smart pointers (ComPtr, unique_ptr)
- Implemented event-driven audio system using IAudioEndpointVolumeCallback to eliminate polling
- Added DPI awareness and dynamic scaling for overlay and recorder windows
- Redesigned Settings UI with a modern tabbed interface (SysTabControl32)
- Improved overall performance and maintainability

+++ src/WasapiRecorder.cpp
@@ -203,7 +203,7 @@ bool WasapiRecorder::SaveToFile(const std::string& filename) {
     if (!file.is_open()) return false;
 
     // Use actual format from pwfx
-    WriteWavHeader(file, audioBuffer.size(), pwfx->nSamplesPerSec, pwfx->nChannels, pwfx->wBitsPerSample);
+    WriteWavHeader(file, (int)audioBuffer.size(), pwfx->nSamplesPerSec, pwfx->nChannels, pwfx->wBitsPerSample);
     
     file.write((char*)audioBuffer.data(), audioBuffer.size());
     file.close();
